% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{rob_get_secret}
\alias{rob_get_secret}
\alias{rob_set_token}
\alias{forget_token}
\title{Get, set and forget authentification token.}
\usage{
rob_get_secret()

rob_set_token(token = NULL)

forget_token()
}
\arguments{
\item{token}{authentification token (obtained at \url{https://owlbot.info/}).}
}
\description{
Get, set and forget authentification token.
}
\section{Functions}{
\itemize{
\item \code{rob_set_token}: set token interactively.

\item \code{forget_token}: forget token.
}}

