% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowlbot.R
\name{rowlbot}
\alias{rowlbot}
\alias{print.rob_def}
\alias{plot.rob_def}
\title{Query the OWLBOT API see \url{https://owlbot.info/}}
\usage{
rowlbot(word)

\method{print}{rob_def}(x, ...)

\method{plot}{rob_def}(x, ...)
}
\arguments{
\item{word}{a word to be defined.}

\item{x}{an object of class \code{rob_def}.}

\item{...}{ignored.}
}
\description{
Query the OWLBOT API see \url{https://owlbot.info/}
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: print method for \code{rob_def} object.

\item \code{plot}: plot method for \code{rob_def} object.
}}

\examples{
\dontrun{
 rowlbot("owl")
}
}
